+++ libutil/op_fileio.c
@@ -95,8 +95,8 @@ void op_read_file(FILE * fp, void * buf, size_t size)
 	if (count != 1) {
 		if (feof(fp)) {
 			fprintf(stderr,
-				"oprofiled:op_read_file: read less than expected %zd bytes\n",
-				size);
+				"oprofiled:op_read_file: read less than expected %lu bytes\n",
+				(unsigned long)size);
 		} else {
 			fprintf(stderr,
 				"oprofiled:op_read_file: error reading\n");
@@ -122,8 +122,8 @@ void op_write_file(FILE * fp, void const * buf, size_t size)
 
 	if (written != 1) {
 		fprintf(stderr,
-			"oprofiled:op_write_file: wrote less than expected: %zd bytes.\n",
-			size);
+			"oprofiled:op_write_file: wrote less than expected: %lu bytes.\n",
+			(unsigned long)size);
 		exit(EXIT_FAILURE);
 	}
 }
